+++ src/core/managers/ToastManager/types/toast.types.ts
@@ -2,6 +2,7 @@ import {
   TransactionBatchStatusesEnum,
   TransactionServerStatusesEnum
 } from 'types/enums.types';
+import { ITransactionsDisplayInfo } from 'types/transactions.types';
 
 export enum TransactionsDefaultTitles {
   success = 'Transaction successful',
@@ -17,6 +18,7 @@ export interface GetToastsOptionsDataPropsType {
   sender: string;
   status?: TransactionBatchStatusesEnum | TransactionServerStatusesEnum;
   toastId: string;
+  transactionsDisplayInfo?: ITransactionsDisplayInfo;
 }
 
 export interface IToastDataState {