+++ src/store/actions/trackedTransactions/trackedTransactionsActions.ts
@@ -4,7 +4,10 @@ import {
   TransactionServerStatusesEnum
 } from 'types/enums.types';
 import { ServerTransactionType } from 'types/serverTransactions.types';
-import { SignedTransactionType } from 'types/transactions.types';
+import {
+  ITransactionsDisplayInfo,
+  SignedTransactionType
+} from 'types/transactions.types';
 import {
   getIsTransactionFailed,
   getIsTransactionNotExecuted,
@@ -20,15 +23,20 @@ export interface UpdateTrackedTransactionStatusPayloadType {
   inTransit?: boolean;
 }
 
-export const createTrackedTransactionsSession = (
-  transactions: SignedTransactionType[]
-) => {
+export const createTrackedTransactionsSession = ({
+  transactions,
+  transactionsDisplayInfo
+}: {
+  transactions: SignedTransactionType[];
+  transactionsDisplayInfo?: ITransactionsDisplayInfo;
+}) => {
   const sessionId = Date.now().toString();
   getStore().setState(
     ({ trackedTransactions: state }) => {
       state[sessionId] = {
         transactions,
-        status: TransactionBatchStatusesEnum.sent
+        status: TransactionBatchStatusesEnum.sent,
+        transactionsDisplayInfo
       };
     },
     false,